updated messaging and fixed tmp file remove issue

+++ db_loader/raw_pop_functions.R
@@ -15,12 +15,15 @@ load_raw_f <- function(
   retry = F) {
 
   file_tmp <- paste0(path_tmptxt, "/tmp.txt")
-  file.remove(file_tmp)
+  
+  if (file.exists(file_tmp)) {
+    file.remove(file_tmp)
+  }
   
   if (grepl(":", file_tmp) == T) { file_load <- gsub("/","\\\\",file_tmp) 
   } else { file_load <- file_tmp }
   
-  message("Writing tmp.txt file")
+  message("...Writing tmp.txt file")
   write.table(data, file = file_tmp, sep = "\t", eol = "\n", quote = F, row.names = F, col.names = T)
   
   create_table_f(conn = conn, config = config, overwrite = !retry)
@@ -29,7 +32,7 @@ load_raw_f <- function(
                      ' -t {config$field_term} -r {config$row_term} -C 65001 -F 2 ',
                      ' -S KCITSQLUTPDBH51 -T -b 100000 -c '))
   
-  message("Loading data into raw.pop")
+  message("...Loading data into raw.pop")
   system2(command = "bcp", args = c(bcp_args))
   
   sql_get <- glue::glue_sql(
@@ -47,7 +50,9 @@ load_raw_f <- function(
       etl_batch_id = rows_loaded[i, 1],
       field = "load_raw_datetime")
   }
-  file.remove(file_tmp)
+  if (file.exists(file_tmp)) {
+    file.remove(file_tmp)
+  }
   return(rows_loaded)
 }
 
@@ -61,7 +66,7 @@ clean_raw_f <- function(
   alter_table_f(conn = conn, config = config)
   
   ### UPDATE FIELDS BASED ON ETL LOG INFO ###
-  message("Updating fields based on ETL Log Info")
+  message("...Updating fields based on ETL Log Info")
   DBI::dbExecute(conn,glue::glue_sql(
     "UPDATE R
     SET R.geo_type = E.geo_type, 
@@ -74,7 +79,7 @@ clean_raw_f <- function(
     .con = conn))
   
   ### FIX RACEMARS TO HAVE LEADING ZEROES ###
-  message("Fixing racemars column")
+  message("...Fixing [racemars] column")
   DBI::dbExecute(conn,glue::glue_sql(
     "UPDATE {`config$schema_name`}.{`config$table_name`}
     SET racemars = RIGHT('0000'+ CAST(racemars AS VARCHAR(5)), 5)
@@ -82,7 +87,7 @@ clean_raw_f <- function(
     .con = conn))
   
   ### FIX AGESTR ###
-  message("Fixing agestr column")
+  message("...Fixing [agestr] column")
   DBI::dbExecute(conn,glue::glue_sql(
     "UPDATE {`config$schema_name`}.{`config$table_name`}
     SET agestr = LEFT(agestr, 3)
@@ -90,7 +95,7 @@ clean_raw_f <- function(
     .con = conn))
   
   ### SET AGE ###
-  message("Setting age column")
+  message("...Setting [age] column")
   DBI::dbExecute(conn,glue::glue_sql(
     "UPDATE {`config$schema_name`}.{`config$table_name`}
     SET age = CAST(agestr AS SMALLINT)
@@ -98,7 +103,7 @@ clean_raw_f <- function(
     .con = conn))
   
   ### SET AGE11 USE REF.POP_CROSSWALK ###
-  message("Setting age11 column")
+  message("...Setting [age11] column")
   DBI::dbExecute(conn,glue::glue_sql(
     "UPDATE R
     SET R.age11 = X.new_value_num
@@ -109,7 +114,7 @@ clean_raw_f <- function(
     .con = conn))
   
   ### SET AGE20 USE REF.POP_CROSSWALK###
-  message("Setting age20 column")
+  message("...Setting [age20] column")
   DBI::dbExecute(conn,glue::glue_sql(
     "UPDATE R
     SET R.age20 = X.new_value_num
@@ -120,7 +125,7 @@ clean_raw_f <- function(
     .con = conn))
   
   ### SET S USE REF.POP_CROSSWALK ###
-  message("Setting s column")
+  message("...Setting [s] column")
   DBI::dbExecute(conn,glue::glue_sql(
     "UPDATE R
     SET R.s = X.new_value_num
@@ -130,7 +135,7 @@ clean_raw_f <- function(
     .con = conn))
   
   ### SET H USE REF.POP_CROSSWALK ###
-  message("Setting h column")
+  message("...Setting [h] column")
   DBI::dbExecute(conn,glue::glue_sql(
     "UPDATE R
     SET R.h = X.new_value_num
@@ -140,7 +145,7 @@ clean_raw_f <- function(
     .con = conn))
   
   ### SET RCODE USE REF.POP_CROSSWALK ###
-  message("Setting rcode column")
+  message("...Setting [rcode] column")
   DBI::dbExecute(conn,glue::glue_sql(
     "UPDATE R
     SET R.rcode = X.new_value_num
@@ -151,7 +156,7 @@ clean_raw_f <- function(
     .con = conn))
   
   ### SET R1_3 USE REF.POP_CROSSWALK ###
-  message("Setting r1_3 column")
+  message("...Setting [r1_3] column")
   DBI::dbExecute(conn,glue::glue_sql(
     "UPDATE R
     SET R.r1_3 = X.new_value_num
@@ -162,7 +167,7 @@ clean_raw_f <- function(
     .con = conn))
   
   ### SET R2_4 USE REF.POP_CROSSWALK ###
-  message("Setting r2_4 column")
+  message("...Setting [r2_4] column")
   DBI::dbExecute(conn,glue::glue_sql(
     "UPDATE R
     SET R.r2_4 = X.new_value_num
@@ -177,7 +182,7 @@ clean_raw_f <- function(
     .con = conn))
   
   ### SET FIPS_CO COLUMN ###
-  message("Setting fips_co column")
+  message("...Setting [fips_co] column")
   DBI::dbExecute(conn,glue::glue_sql(
     "UPDATE {`config$schema_name`}.{`{config$table_name}`}
     SET fips_co = CAST(SUBSTRING(geo_id, 3, 3) AS SMALLINT)
